them vao giỏ hàng

+++ src/main/java/NgoGiaSam/Web_Elaban_be/dao/UserRespository.java
@@ -6,7 +6,7 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 
 @RepositoryRestResource(path = "users")
-public interface UserRespository extends JpaRepository<User,Integer> {
+public interface UserRespository extends JpaRepository<User,Long> {
     public   boolean existsByUsername(String username);
 
     public   boolean existsByEmail(String email);
